fix: add text trunkate [MDS-578]
