+++ workspaces/core/src/selectButton/private/types/ChipProps.ts
@@ -1,0 +1,3 @@
+type ChipProps = { children: React.ReactNode; onClear?: () => void };
+
+export default ChipProps;
