+++ workspaces/core/src/selectButton/private/types/InputProps.ts
@@ -1,0 +1,6 @@
+type InputProps = {
+  className?: string;
+  children?: React.ReactNode;
+};
+
+export default InputProps;
