+++ workspaces/core/src/selectButton/private/types/LabelProps.ts
@@ -1,0 +1,7 @@
+type LabelProps = {
+  labelSize?: 'sm' | 'md' | 'lg' | 'xl' | string;
+  idDisabled?: boolean;
+  children?: React.ReactNode;
+};
+
+export default LabelProps;
