+++ workspaces/core/src/selectButton/private/types/SelectButtonProps.ts
@@ -1,0 +1,11 @@
+type SelectButtonProps = {
+  size?: 'sm' | 'md' | 'lg' | string;
+  isError?: boolean;
+  open?: boolean;
+  idDisabled?: boolean;
+  label?: JSX.Element | string;
+  placeholder?: JSX.Element | string;
+  children?: React.ReactNode;
+};
+
+export default SelectButtonProps;
