+++ workspaces/core/src/selectButton/private/types/SelectButtonState.ts
@@ -1,0 +1,8 @@
+type SelectButtonState = {
+  open?: boolean;
+  size?: 'sm' | 'md' | 'lg' | string;
+  isError?: boolean;
+  idDisabled?: boolean;
+};
+
+export default SelectButtonState;
