+++ workspaces/core/src/selectButton/private/utils/SelectButtonContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import type SelectButtonState from '../types/SelectButtonState';
+
+const SelectButtonContext = createContext<SelectButtonState>({});
+SelectButtonContext.displayName = 'InputBtnContext';
+
+export default SelectButtonContext;
